feat: add support for v2
* Modified models to fit the new data model used in v2
* Added support for conneciton-based comm mode
* Added animation models

+++ lib/models/device/list_of_device_play.dart
@@ -11,7 +11,7 @@ class ListOfDevicePlayModels extends ValueNotifier<List<DevicePlayModel>> {
     notifyListeners();
   }
 
-  void updateCapState(double newCapState, int index) {
+  void updateCapState(int newCapState, int index) {
     if (index < 0 || index >= value.length) return;
     value[index].updateCapState(newCapState);
     notifyListeners();
